First commited as before_84527_rev9

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/file/GC.java
@@ -459,9 +459,14 @@ public class GC {
 			return;
 
 		// delete all candidates which have survived: these are unreferenced
-		// loose objects
-		for (File f : deletionCandidates.values())
-			f.delete();
+		// loose objects. Make a last check, though, to avoid deleting objects
+		// that could have been referenced while the candidates list was being
+		// built (by an incoming push, for example).
+		for (File f : deletionCandidates.values()) {
+			if (f.lastModified() < expireDate) {
+				f.delete();
+			}
+		}
 
 		repo.getObjectDatabase().close();
 	}
@@ -602,7 +607,7 @@ public class GC {
 			nonHeads.addAll(listRefLogObjects(ref, 0));
 			if (ref.isSymbolic() || ref.getObjectId() == null)
 				continue;
-			if (ref.getName().startsWith(Constants.R_HEADS))
+			if (isHead(ref) || isTag(ref))
 				allHeads.add(ref.getObjectId());
 			else if (RefTreeNames.isRefTree(refdb, ref.getName()))
 				txnHeads.add(ref.getObjectId());
@@ -653,6 +658,14 @@ public class GC {
 		lastPackedRefs = refsBefore;
 		lastRepackTime = time;
 		return ret;
+	}
+
+	private static boolean isHead(Ref ref) {
+		return ref.getName().startsWith(Constants.R_HEADS);
+	}
+
+	private static boolean isTag(Ref ref) {
+		return ref.getName().startsWith(Constants.R_TAGS);
 	}
 
 	/**
@@ -1199,9 +1212,10 @@ public class GC {
 				new DirectoryStream.Filter<Path>() {
 
 					public boolean accept(Path file) throws IOException {
-						return Files.isRegularFile(file) && PATTERN_LOOSE_OBJECT
-								.matcher(file.getFileName().toString())
-								.matches();
+						Path fileName = file.getFileName();
+						return Files.isRegularFile(file) && fileName != null
+								&& PATTERN_LOOSE_OBJECT
+										.matcher(fileName.toString()).matches();
 					}
 				})) {
 			for (Iterator<Path> iter = stream.iterator(); iter.hasNext();
